@@ -16,10 +16,10 @@
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.2.0",
     "@vitejs/plugin-vue": "^4.0.0",
-    "eslint": "^8.34.0",
-    "eslint-plugin-vue": "^9.9.0",
+    "eslint": "^9.20.1",
+    "eslint-plugin-vue": "^9.32.0",
     "sass": "^1.58.2",
-    "vite": "^4.1.2",
+    "vite": "6.0.9",
     "vite-plugin-inspect": "^0.7.15",
     "stylelint": "^14.16.1",
     "stylelint-config-prettier": "^9.0.4",
