@@ -1,45 +1,47 @@
 <template>
-  <form
-    @submit.prevent="changeUserEmail"
-  >
-    <div v-if="errors.length" class="errorsInfo">
-      <b>Пожалуйста исправьте указанные ошибки:</b>
-      <ul>
-        <li
-          v-for="error in errors"
-          :key="error"
-        >
-          {{ error }}
-        </li>
-      </ul>
-    </div>
-    <div class="wrapper-change-email">
-      <h2>Смена почты</h2>
-      <my-input
-        v-model.trim="user.newEmail"
-        placeholder="Новый email:"
-        type="text"
-      />
-      <my-input
-        v-model.trim="user.repeatedNewEmail"
-        placeholder="Подтвердите email:"
-        type="text"
-      />
-      <my-input
-        v-model.trim="user.password"
-        placeholder="Пароль:"
-        type="password"
-      />
-      <button
-        class="submit-btn"
-        type="submit"
-      >Сменить почту  
-      </button>
-    </div>
-    </form>
+	<form
+		@submit.prevent="changeUserEmail"
+	>
+	    <div v-if="errors.length" class="errorsInfo">
+	        <b>Пожалуйста исправьте указанные ошибки:</b>
+	        <ul>
+		        <li
+		            v-for="error in errors"
+		            :key="error"
+		        >
+		            {{ error }}
+		        </li>
+	        </ul>
+	    </div>
+	    <div class="wrapper-change-email">
+	        <h2>Смена почты</h2>
+	        <my-input
+		        v-model.trim="user.value.newEmail"
+		        placeholder="Новый email:"
+		        type="text"
+	         />
+	        <my-input
+		        v-model.trim="user.value.repeatedNewEmail"
+		        placeholder="Подтвердите email:"
+		        type="text"
+	        />
+	        <my-input
+		        v-model.trim="user.value.password"
+		        placeholder="Пароль:"
+		        type="password"
+	        />
+	        <my-button
+		        class="submit-btn"
+		        type="submit"
+	        >
+		        Сменить почту
+	        </my-button>
+	    </div>
+	</form>
 </template>
 
-<script>
+<script setup>
+import { ref, defineEmits } from 'vue'
 import MyButton from '@/components/UI/MyButton.vue'
 import MyInput from './UI/MyInput.vue';
 import { getAuth, updateEmail, signOut } from 'firebase/auth'
@@ -47,87 +49,66 @@ import { checkOfRegistration } from '@/stores/checkOfRegistration'
 import { doc, updateDoc } from 'firebase/firestore'
 import { db } from '@/main'
 
-export default {
-  name: 'change-email',
+const userStore = checkOfRegistration()
+const user = ref( {
+    newEmail: '',
+    repeatedNewEmail: '',
+    password: ''
+  })
+const errors = ref([])
+const auth = getAuth()
+const emit = defineEmits(['hide'])
+const changeUserEmail = async()=>  {
+	if ( user.value.newEmail === user.value.repeatedNewEmail) {
+		updateEmail(auth.currentUser, user.value.newEmail)
+		.catch((error) => {
+			console.log(error);
+		})
 
-  components: {
-    MyButton,
-    MyInput
-  },
-  setup: () => ({
-    userStore: checkOfRegistration(),
-  }),
-  data() {
-    return {
-      user: {
-        newEmail: '',
-        repeatedNewEmail: '',
-        password: ''
-      },
-      errors: []
+		console.log(auth.currentuser.value.uid);
+
+		await updateDoc(doc(db, 'users', auth.currentuser.value.uid), {
+			email: user.value.newEmail
+		})
+		.catch((error) => {
+		    console.log(error);
+		})
+
+		localStorage.clear()
+		userStore.$reset()
+		emit('hide')
+		window.location.reload();
     }
-  },
-
-  methods: {
-    async changeUserEmail() {
-      if ( this.user.newEmail === this.user.repeatedNewEmail) {
-          // samorod1010@mail.ru
-          const auth = getAuth()
-          updateEmail(auth.currentUser, this.user.newEmail).then(()=> {            
-            console.log('email updated!');
-          })
-          .catch(() => {
-            console.log(error);
-          })
-
-          console.log(auth.currentUser.uid);
-          
-          await updateDoc(doc(db, 'users', auth.currentUser.uid), {
-              email: this.user.newEmail
-          })
-          .catch(() => {
-                console.log(error);
-          })
-
-          localStorage.clear()
-          this.userStore.$reset()
-          this.$emit('hide')
-          window.location.reload();
-        } 
-        
-        else {
-            alert('Почты не совпадают!')
-        }
-        if (this.user.newEmail === '') {
-        this.errors.push('Введите почту!')
-        }
-        if (this.user.password !== this.user.password) {
-        this.errors.push('Неправильный пароль!')
-        }
-        if (this.user.newEmail === '' || this.user.newEmail !== this.user.repeatedNewEmail || this.user.repeatedNewEmail === '') {
-            this.errors.push('Почты не совпадают или не заполнены поля!')
-        }
+    else {
+        alert('Почты не совпадают!')
     }
-  }
+    if (user.value.newEmail === '') {
+        errors.value.push('Введите почту!')
+    }
+    if (user.value.password !== user.value.password) {
+        errors.value.push('Неправильный пароль!')
+    }
+    if (user.value.newEmail === '' || user.value.newEmail !== user.value.repeatedNewEmail || user.value.repeatedNewEmail === '') {
+        errors.value.push('Почты не совпадают или не заполнены поля!')
+    }
 }
 </script>
 
 <style scoped>
 
 .wrapper-change-email {
-  display: grid;
+	display: grid;
 }
 
 .btn-account {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 300px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	width: 300px;
 }
 
 .submit-btn {
     margin-top: 20px;
     cursor: pointer;
-  }
+}
 </style>
